personalpyy/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pandas
plotly
requests
streamlit
